requirements.txt
eventlet==0.25.2
Flask==1.1.2
flask-socketio==4.3.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Markupsafe==1.1.1
redis-py==3.5.3
Werkzeug==1.0.1
